+++ packages/g-lite/src/css/properties/CSSPropertyPath.ts
@@ -48,6 +48,9 @@ export class CSSPropertyPath
    * update local position
    */
   postProcessor(object: DisplayObject, attributes: string[]) {
+    object.parsedStyle.defX = object.parsedStyle.path.rect.x;
+    object.parsedStyle.defY = object.parsedStyle.path.rect.y;
+
     if (
       object.nodeName === Shape.PATH &&
       attributes.indexOf('transform') === -1