+++ site/examples/canvas/container/demo/supports-css-transform.js
@@ -14,7 +14,9 @@ $wrapper.style.transform = 'scale(1.1)';
 const canvasRenderer = new CanvasRenderer();
 const svgRenderer = new SVGRenderer();
 const webglRenderer = new WebGLRenderer();
-const webgpuRenderer = new WebGPURenderer();
+const webgpuRenderer = new WebGPURenderer({
+  shaderCompilerPath: '/glsl_wgsl_compiler_bg.wasm',
+});
 const canvaskitRenderer = new CanvaskitRenderer({
   wasmDir: '/',
 });