+++ site/examples/plugins/zdog/demo/zdog.js
@@ -17,7 +17,7 @@ const canvas = new Canvas({
   container: 'container',
   width: 600,
   height: 500,
-  renderer: svgRenderer,
+  renderer: canvasRenderer,
 });
 
 canvas.addEventListener(CanvasEvent.READY, () => {