+++ packages/g-dom-mutation-observer-api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @antv/g-dom-mutation-observer-api
 
+## 1.2.2
+
+### Patch Changes
+
+-   Updated dependencies [0eb5142d]
+-   Updated dependencies [71990540]
+    -   @antv/g-lite@1.2.2
+
 ## 1.2.1
 
 ### Patch Changes