+++ packages/g-mobile-canvas-element/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @antv/g-mobile-canvas-element
 
+## 0.8.2
+
+### Patch Changes
+
+-   Updated dependencies [0eb5142d]
+-   Updated dependencies [71990540]
+    -   @antv/g-lite@1.2.2
+
 ## 0.8.1
 
 ### Patch Changes