+++ packages/g-svg/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @antv/g-svg
 
+## 1.10.2
+
+### Patch Changes
+
+-   Updated dependencies [0eb5142d]
+-   Updated dependencies [71990540]
+    -   @antv/g-lite@1.2.2
+    -   @antv/g-plugin-dom-interaction@1.9.2
+    -   @antv/g-plugin-svg-picker@1.9.2
+    -   @antv/g-plugin-svg-renderer@1.10.2
+
 ## 1.10.1
 
 ### Patch Changes